// Capstone Project/src/app/pages/admin/admin-transaction-management-page/screen/types.ts
export enum adminTransactionPageEnum {
  ADMIN_TRANSACTION_REQUEST = 'APP/ADMIN_TRANSACTION/ADMIN_TRANSACTION_REQUEST',
  ADMIN_TRANSACTION_SUCCESS = 'APP/ADMIN_TRANSACTION/ADMIN_TRANSACTION_SUCCESS',
  ADMIN_TRANSACTION_FAILURE = 'APP/ADMIN_TRANSACTION/ADMIN_TRANSACTION_FAILURE',

  ADMIN_TRANSACTION_CONFIRM_REQUEST = 'APP/ADMIN_TRANSACTION/ADMIN_TRANSACTION_CONFIRM_REQUEST',
  ADMIN_TRANSACTION_CONFIRM_SUCCESS = 'APP/ADMIN_TRANSACTION/ADMIN_TRANSACTION_CONFIRM_SUCCESS',
  ADMIN_TRANSACTION_CONFIRM_FAILURE = 'APP/ADMIN_TRANSACTION/ADMIN_TRANSACTION_CONFIRM_FAILURE',

  ADMIN_TRANSACTION_REJECT_REQUEST = 'APP/ADMIN_TRANSACTION/ADMIN_TRANSACTION_REJECT_REQUEST',
  ADMIN_TRANSACTION_REJECT_SUCCESS = 'APP/ADMIN_TRANSACTION/ADMIN_TRANSACTION_REJECT_SUCCESS',
  ADMIN_TRANSACTION_REJECT_FAILURE = 'APP/ADMIN_TRANSACTION/ADMIN_TRANSACTION_REJECT_FAILURE',

  ADMIN_TRANSACTION_SEARCH_REQUEST = 'APP/ADMIN_TRANSACTION/ADMIN_TRANSACTION_SEARCH_REQUEST',
  ADMIN_TRANSACTION_SEARCH_SUCCESS = 'APP/ADMIN_TRANSACTION/ADMIN_TRANSACTION_SEARCH_SUCCESS',
  ADMIN_TRANSACTION_SEARCH_FAILURE = 'APP/ADMIN_TRANSACTION/ADMIN_TRANSACTION_SEARCH_FAILURE',

  ADMIN_TRANSACTION_CREATE_REQUEST = 'APP/ADMIN_TRANSACTION/ADMIN_TRANSACTION_CREATE_REQUEST',
  ADMIN_TRANSACTION_CREATE_SUCCESS = 'APP/ADMIN_TRANSACTION/ADMIN_TRANSACTION_CREATE_SUCCESS',
  ADMIN_TRANSACTION_CREATE_FAILURE = 'APP/ADMIN_TRANSACTION/ADMIN_TRANSACTION_CREATE_FAILURE',

  ADMIN_TRANSACTION_POST_OR_EXTEND_SEARCH_REQUEST = 'APP/ADMIN_TRANSACTION/ADMIN_TRANSACTION_POST_OR_EXTEND_SEARCH_REQUEST',
  ADMIN_TRANSACTION_POST_OR_EXTEND_SEARCH_SUCCESS = 'APP/ADMIN_TRANSACTION/ADMIN_TRANSACTION_POST_OR_EXTEND_SEARCH_SUCCESS',
  ADMIN_TRANSACTION_POST_OR_EXTEND_SEARCH_FAILURE = 'APP/ADMIN_TRANSACTION/ADMIN_TRANSACTION_POST_OR_EXTEND_SEARCH_FAILURE',

  CLEAR_STATE = 'APP/ADMIN_TRANSACTION/CLEAR_STATE ',
}